cert-manager: change resource type of Issuer to ClusterIssuer
According to https://cert-manager.io/docs/concepts/issuer/#namespaces

An Issuer is a namespaced resource, and it is not possible to issue
certificates from an Issuer in a different namespace. This means you will
need to create an Issuer in each namespace you wish to obtain Certificates
in.

If you want to create a single Issuer that can be consumed in multiple
namespaces, you should consider creating a ClusterIssuer resource. This is
almost identical to the Issuer resource, however is non-namespaced so it
can be used to issue Certificates across all namespaces.