home-assistant: add annotation ingress.class: public
According to the annotations documentation:
https://cert-manager.io/docs/usage/ingress/#supported-annotations

This annotation can be used to have a different ingress for acme than with
the ingress for which the certificate is generated:
`acme.cert-manager.io/http01-ingress-class`

However, it also implies that it should use the same ingress class as the
ingress for which the certificate is generated.

It seems like the class is "public" rather than "nginx":
$ kubectl -n ingress describe daemonset.apps/nginx-ingress-microk8s-controller | grep class
      --ingress-class=public

Try specifying the ingress for home-assistant to see if acme will work.