librespeed: switch back to http ingress
Uploads work through the http ingress but not https.

Note that the issue last time this was tried was that chrome
preserved the HSTS set up by the TLS ingress for https. The
upload test worked fine from a guest user, but my personal
user was always forced to https. Clearing HSTS and cache was
required to get it working.

This reverts commit 0f5f9615c6820625338969177b00312474da1e2e.